feat: :white_check_mark: Add password, login page, and update user model
create a password field on the register page create a login page with username and password field. handle the submit action on login page. It should redirect to the user page when to login is successful. Handle all error cases (on front-end and back-end side) for register and login page Add firstname and lastname field to user model. Make them as required in the form update user page to display